add mass pvp, fixed bugs

+++ Assets/Scripts/Fighter/Weapon.cs
@@ -8,7 +8,7 @@ public class Weapon : ScriptableObject
     [SerializeField] private string _name;
     [SerializeField][Range(1,100)] private int _minDamage;
     [SerializeField][Range(2, 200)] private int _maxDamage;
-    [SerializeField][Range(1, 10)] private int _atackSpeed;
+    [SerializeField][Range(1, 10)] private float _atackSpeed;
     [SerializeField][Range(1, 20)] private int _criticalChance;
     [SerializeField][Range(1.5f, 5)] private float _criticalModifier;
     [SerializeField][Range(5, 20)] private int _missChance;
@@ -18,7 +18,7 @@ public class Weapon : ScriptableObject
     public string Name => _name;
     public int MinDamage => _minDamage;
     public int MaxDamage => _maxDamage;
-    public int AtackSpeed => _atackSpeed;
+    public float AtackSpeed => _atackSpeed;
     public int CriticalChance => _criticalChance;
     public float CriticalModifier => _criticalModifier;
     public int MissChance => _missChance;
